requirements.txt
# requirements.txt
folium
Jinja2
Markdown
python-frontmatter
beautifulsoup4
# Aggiungi altre dipendenze se ne usi (es. requests se carichi dati esterni)
